refactor(core): overhaul Dictation architecture and establish domain boundaries This massive refactoring strictly enforces single-responsibility and domain isolation across the macroscopic Audio, Transcription, Rewrite, and SDK networking layers. Key achievements: 1. Audio Decoupling: Extracted . Renamed  to , confining it purely to microphone interactions. 2. Pipeline Factory: Extracted dependency and execution routing into a dedicated , slimming down the  to a lightweight coordinator. 3. Domain Isolation: Created  and  directories. Decoupled Audio-to-Text from Text-to-Text operations through generic modular protocols. 4. SDK Cleanup: Demolished the monolithic [OpenAISDKSupport.swift](cci:7://file:///Users/nd/Developer/Stet/apps/mac/Stet/Core/OpenAI/OpenAISDKSupport.swift:0:0-0:0) file into 7 strict SDK configuration helpers, and completely purged dead historical middleware logs. Validates with Exit Code 0 on macOS arm64 xcodebuild. No UI contract breakages.
