+++ apps/mac/Stet/Core/Speech/AudioCaptureService.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+protocol AudioCaptureService: Sendable {
+    func startRecording() async throws
+    func stopRecording() async throws -> (url: URL, duration: TimeInterval?)
+    func cancelRecording() async
+}
