+++ apps/mac/Stet/Core/Transcribed/AudioFileTranscriptionService.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+protocol AudioFileTranscriptionService: Sendable {
+    func transcribe(
+        audioFileAt fileURL: URL,
+        languageCode: String?,
+        prompt: String?,
+        audioDurationSeconds: TimeInterval?
+    ) async throws -> String
+}
