+++ apps/mac/Stet/Core/Transcribed/OpenAITranscriptionService.swift
@@ -1,15 +1,6 @@
 import Foundation
 import OpenAI
 
-protocol AudioFileTranscriptionService: Sendable {
-    func transcribe(
-        audioFileAt fileURL: URL,
-        languageCode: String?,
-        prompt: String?,
-        audioDurationSeconds: TimeInterval?
-    ) async throws -> String
-}
-
 struct OpenAITranscriptionService: AudioFileTranscriptionService {
     private struct APIErrorEnvelope: Decodable {
         let error: APIErrorPayload
