+++ apps/mac/Stet/Core/Transcribed/TranscriptionPreparation.swift
