+++ apps/mac/Stet/Core/Transcribed/TranscriptionUploadAudioFormat.swift
